coder

+++ scripts/main.js
@@ -1,9 +1,9 @@
 // Seleccionamos moneda a utilizar, en este caso pesos argentinos
 const formatter = new Intl.NumberFormat('es-AR',{style: 'currency', currency: 'ARS'})
 
-// Definimos saldo inicial y manipulamos DOM
+// Definimos saldo líquido inicial y manipulamos DOM
 const initialValue = 100000;
-if (localStorage.getItem('saldo')===100000) {
+if (localStorage.getItem('saldo')===100000 || NaN) {
     localStorage.setItem('saldo', initialValue)
 }
 
@@ -12,10 +12,10 @@ for (let item of saldo) {
     console.log(item.innerHTML= formatter.format((localStorage.getItem("saldo"))));
 }
 
-// Definimos saldo Inversiones inicial y manipulamos DOM
+// Definimos saldo Inversiones inicial 0 y manipulamos DOM
 const initialValueInversiones = 0;
-if (localStorage.getItem('inversiones')===0) {
-    localStorage.setItem('inversiones', initialValue)
+if (localStorage.getItem('inversiones')===0 || NaN) {
+    localStorage.setItem('inversiones', initialValueInversiones)
 }
 
 const inversiones =  document.getElementsByClassName('inversiones');
@@ -37,26 +37,30 @@ const calcularTotal = (precio, cantidad, comisiones) => {
     let calcularCometa = calcularSubtotal * comisiones;
     let saldoAImprimir = localStorage.getItem('saldo');
     const total = calcularSubtotal + calcularCometa;
-    
-    if (total<initialValue) {
-        let calculoFinal = saldoAImprimir - total;
-        console.log(calculoFinal)
-         if (typeof localStorage.getItem('saldo')!=='undefined'){
-             localStorage.setItem('saldo', (localStorage.getItem('saldo') - total))
-          }else{
-              localStorage.setItem('saldo', (calculoFinal))
-            }
-
-         if (typeof localStorage.getItem('inversiones')!=='undefined'){
-             localStorage.setItem('inversiones', total)
-          }else{
-              localStorage.setItem('inversiones', (total))
-            }
-        return formatter.format(total)
-    }else{
-        alert('No posee suficiente saldo, revisar importes ingresados')
+    if (confirm('Confirmar orden')) {
+        if (total<initialValue) {
+            let calculoFinal = saldoAImprimir - total;
+             if (typeof localStorage.getItem('saldo')!=='undefined'){
+                 localStorage.setItem('saldo', (localStorage.getItem('saldo') - total))
+              }else{
+                  localStorage.setItem('saldo', calculoFinal)
+                }
+             if (typeof localStorage.getItem('inversiones')!=='undefined'){
+                 localStorage.setItem('inversiones', total + Number.parseFloat(localStorage.getItem('inversiones')) )
+              }else{
+                  localStorage.setItem('inversiones', total + Number.parseFloat(localStorage.getItem('inversiones')) )
+                }
+                document.getElementById('total1').innerHTML = "Total: " + formatter.format(total);
+                document.getElementById('confirmed-order').classList.add('show')
+                document.getElementById('card-body').classList.remove('show')
+                document.getElementById('card-body').classList.add('collapse')
+            return formatter.format(total)
+        }else{
+            document.getElementById('denied-order').classList.add('show')
+            document.getElementById('card-body').classList.remove('show')
+            document.getElementById('card-body').classList.add('collapse')
+        }
     }
-    
 }
 
 const capturarOnClick = () => {
@@ -69,10 +73,16 @@ const capturarOnClick = () => {
 
     // Guardo string en localStorage
     const orden = {ordenes: [{producto: miPedido.producto, precio: miPedido.precio, cantidad: miPedido.cantidad, comisiones: miPedido.comisiones, total: miPedido.total}]}
-    if (!miPedido.producto || !miPedido.precio || !miPedido.cantidad){
+
+    if (!miPedido.producto || !miPedido.precio || !miPedido.cantidad ){
         alert('Error en la orden, falta completar uno o más campos')
         return
+    }else{
+        if (miPedido.total == undefined) {
+            return
+        }
     }
+
     // Obtengo string y se pasa a objeto
     let ordenes = JSON.parse(localStorage.getItem('ordenes'))
     if (ordenes) {
@@ -94,6 +104,51 @@ jquery.ordenes.forEach(item => {
         <li>${item.precio}
         <li>${item.cantidad}
         <li class='text-success'>${'Operada'}
+        <li>${item.total}
+        </ul>`
+     );
+});
+
+const jquery2 = JSON.parse(localStorage.getItem('ordenes'))
+
+const arrayDeValores = jquery2.ordenes.map(function(orden){
+    return orden.producto
+})
+
+console.log(arrayDeValores)
+
+let modifiedArray = arrayDeValores.reduce((acc, curr) => {
+    if (!acc.length) {
+      acc.push([curr, 1]);
+    } else {
+      let lastPushedArray = acc[acc.length-1];
+      if (lastPushedArray[0] === curr) {
+        acc[acc.length-1][1]++;
+      } else {
+        acc.push([curr, 1]);
+      }
+    }
+    return acc;
+  }, []);
+  
+  console.log(modifiedArray);
+
+// for (let i = 0; i < modifiedArray.length; i++) {
+//     const element = modifiedArray[i];
+//     for (let index = 0; index < element.length; index++) {
+//         const element2 = element[index];
+//         console.log(element2)
+//     }
+// }
+
+// Utilizamos jQuery para inyectar órdenes creadas
+jquery2.ordenes.forEach(item => {
+    $('.portfolio').prepend(
+        `
+        <ul class="account-main-list">
+        <li>${item.cantidad}
+        <li>${item.producto}
+        <li>${item.precio}
         <li>${item.total}
         </ul>`
      );
